fix: ubuntu.sh duplicate python build dependencies + only install flatpak when we have an x/wayland session runnign

+++ installers/ubuntu.sh
@@ -7,27 +7,23 @@ cd "$DIR" || exit 1
 
 sudo apt-get update
 
-# Python Build
-sudo apt update
-sudo apt install build-essential libssl-dev zlib1g-dev \
-  libbz2-dev libreadline-dev libsqlite3-dev curl git \
-  libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev -y
-
 sudo apt-get install apt-file -y
 
 # XDG utils is needed for gcloud for example
 sudo apt-get install -y vim git curl zsh flatpak build-essential \
-  xdg-utils flatpak-xdg-utils unzip python3-pip python3-venv composer
+  xdg-utils flatpak-xdg-utils unzip python3-pip python3-venv
 
-sudo snap remove firefox
-sudo flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
+if [[ "$DISPLAY" != '' ]]; then
+  sudo snap remove firefox
+  sudo flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
 
-sudo flatpak install flathub org.mozilla.firefox -y
+  sudo flatpak install flathub org.mozilla.firefox -y
+fi
 #
 # dependencies for building Python
 sudo apt-get install -y build-essential libssl-dev zlib1g-dev \
   libbz2-dev libreadline-dev libsqlite3-dev curl git \
-  libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev
+  libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev -y
 # TODO: only if we have gnome installed
 # sudo apt-get install -y gnome-tweaks
